Support mmap file handler for Win32. Thanks to Alan C. L. Sung's help of his experiences on YASA (Yet Another Suffix Array).

+++ Modules/SharedSource/OVCIN.cpp
@@ -107,7 +107,7 @@ void OVCIN::parseCinVector(const vector<string>& cinVector){
         string::size_type del_pos;
         if( (del_pos=line.find_first_of(delimiters)) != string::npos ){
             string key = line.substr(0, del_pos);
-            unsigned int value_pos=line.find_first_not_of(delimiters, del_pos);
+            unsigned int value_pos=(unsigned int)line.find_first_not_of(delimiters, del_pos);
             if( value_pos == string::npos )
                 continue;
             string value = line.substr( value_pos, line.length() - value_pos );
@@ -216,7 +216,7 @@ int OVCIN::findClosestLowerBound(const CinMap& m, const string& key) const{
 
 
 // discarded test routines
-
+/*
 void OVCIN::show(const CinMap &m, int x) {
     if (x==-1) {
         printf ("    not found\n");
@@ -269,5 +269,5 @@ void OVCIN::test() {
     runtest("zxcz");
     runtest("zzzz");    
 }
-
+*/
